fix: Fix logic for count_treasures
count_treasures function will now correctly calc
the number of adjacent treasures and traps. I
utilized a different approach than what I
initially planned.

Instead of creating a list slice up to a k+ or -1
to the player's coords, I used a seperate list for
direction, and added those values to a new x and y
coord for the player to indicate the search
direction. Afterwards, I checked each indexed
location on the map to check for 'T' or 'X'.